+++ Classes/EqualizerViewController.m
@@ -553,7 +553,7 @@
 	settingsS.gainMultiplier = gainValue;
 	audioEngineS.equalizer.gain = gainValue;
 	
-	CGFloat difference = fabsf(gainValue - self.lastGainValue);
+	CGFloat difference = fabs(gainValue - self.lastGainValue);
 	if (difference >= .1 || gainValue == minValue || gainValue == maxValue)
 	{
 		self.gainBoostAmountLabel.text = [NSString stringWithFormat:@"%.1fx", gainValue];
